SDK-1541: Move check report classes
