+++ src/doc_scan/index.js
@@ -2,8 +2,8 @@ const SandboxDocScanClientBuilder = require('./sandbox.doc.scan.client.builder')
 const SandboxCheckReportsBuilder = require('./request/sandbox.check.reports.builder');
 const SandboxTaskResultsBuilder = require('./request/sandbox.task.results.builder');
 const SandboxResponseConfigBuilder = require('./request/sandbox.response.config.builder');
-const SandboxBreakdownBuilder = require('./request/sandbox.breakdown.builder');
-const SandboxRecommendationBuilder = require('./request/sandbox.recommendation.builder');
+const SandboxBreakdownBuilder = require('./request/check/report/sandbox.breakdown.builder');
+const SandboxRecommendationBuilder = require('./request/check/report/sandbox.recommendation.builder');
 const SandboxDocumentAuthenticityCheckBuilder = require('./request/check/sandbox.document.authenticity.check.builder');
 const SandboxDocumentFaceMatchCheckBuilder = require('./request/check/sandbox.document.face.match.check.builder');
 const SandboxZoomLivenessCheckBuilder = require('./request/check/sandbox.zoom.liveness.check.builder');
