+++ src/doc_scan/request/check/report/sandbox.breakdown.builder.js
@@ -1,4 +1,4 @@
-const { Validation } = require('../../util');
+const { Validation } = require('../../../../util');
 const SandboxBreakdown = require('./sandbox.breakdown');
 const SandboxDetail = require('./sandbox.detail');
