+++ src/doc_scan/request/check/report/sandbox.breakdown.js
@@ -1,4 +1,4 @@
-const { Validation } = require('../../util');
+const { Validation } = require('../../../../util');
 const SandboxDetail = require('./sandbox.detail');
 
 class SandboxBreakdown {
